Bump to Tree-sitter 0.26 (#1149)
* chore(deps): update tree-sitter packages

* Remove once-deprecated-now-removed API wrappers

* Fix unnecessary casts

* More removed stuff

* Migrate to new tree_sitter_loader error type

* Bump version to v0.7.3

* Update Cargo.lock

* Update CHANGELOG for release today

---------

+++ topiary-tree-sitter-facade/src/node.rs
@@ -22,7 +22,7 @@ mod native {
 
         #[inline]
         pub fn child(&self, i: u32) -> Option<Self> {
-            self.inner.child(i as usize).map(Into::into)
+            self.inner.child(i).map(Into::into)
         }
 
         #[inline]
@@ -156,7 +156,7 @@ mod native {
 
         #[inline]
         pub fn named_child(&self, i: u32) -> Option<Self> {
-            self.inner.named_child(i as usize).map(Into::into)
+            self.inner.named_child(i).map(Into::into)
         }
 
         #[inline]
